Updates to make the app conversational and cognitive
Use Watson Conversation to provide a natural language interface
to the app, and identify intents, actions, and city entities.

Implement a multi-turn conversational interface, ask the user
for clarifications and confirmations, and remember the state
of the conversation across multiple message exchanges.

Switch from the Weather Underground API to the Weather Company Data
service from Bluemix.